--- src/main/java/com/promptoven/exhibitionservice/admin/banner/application/AdminExhibitionService.java
@@ -1,14 +1,0 @@
-package com.promptoven.exhibitionservice.admin.banner.application;
-
-import com.promptoven.exhibitionservice.admin.banner.dto.in.AddExhibitionRequestDto;
-import com.promptoven.exhibitionservice.admin.banner.dto.in.DeleteExhibitionRequestDto;
-import com.promptoven.exhibitionservice.admin.banner.dto.in.UpdateExhibitionRequestDto;
-
-public interface AdminExhibitionService {
-
-    void addExhibition(AddExhibitionRequestDto exhibitionRequestDto);
-
-    void updateExhibition(UpdateExhibitionRequestDto exhibitionRequestDto);
-
-    void deleteExhibition(DeleteExhibitionRequestDto exhibitionRequestDto);
-}
