+++ src/main/java/com/promptoven/exhibitionservice/admin/exhibition/dto/in/AddExhibitionRequestDto.java
@@ -1,6 +1,6 @@
-package com.promptoven.exhibitionservice.admin.banner.dto.in;
+package com.promptoven.exhibitionservice.admin.exhibition.dto.in;
 
-import com.promptoven.exhibitionservice.admin.banner.vo.in.UpdateExhibitionRequestVo;
+import com.promptoven.exhibitionservice.admin.exhibition.vo.in.AddExhibitionRequestVo;
 import com.promptoven.exhibitionservice.common.domain.Exhibition;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,9 +10,7 @@ import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor
-public class UpdateExhibitionRequestDto {
-
-    private Long exhibitionId;
+public class AddExhibitionRequestDto {
     private String name;
     private String description;
     private String rewardType;
@@ -23,10 +21,9 @@ public class UpdateExhibitionRequestDto {
     private LocalDateTime endDate;
 
     @Builder
-    public UpdateExhibitionRequestDto(Long exhibitionId, String name, String description, String rewardType,
-                                      LocalDateTime bannerStartDate, LocalDateTime bannerEndDate, boolean willShow,
-                                      LocalDateTime startDate, LocalDateTime endDate) {
-        this.exhibitionId = exhibitionId;
+    public AddExhibitionRequestDto(String name, String description, String rewardType,
+                                   LocalDateTime bannerStartDate, LocalDateTime bannerEndDate, boolean willShow,
+                                   LocalDateTime startDate, LocalDateTime endDate) {
         this.name = name;
         this.description = description;
         this.rewardType = rewardType;
@@ -37,23 +34,21 @@ public class UpdateExhibitionRequestDto {
         this.endDate = endDate;
     }
 
-    public static UpdateExhibitionRequestDto toDto(UpdateExhibitionRequestVo vo) {
-        return UpdateExhibitionRequestDto.builder()
-                .exhibitionId(vo.getExhibitionId())
-                .name(vo.getName())
-                .description(vo.getDescription())
-                .rewardType(vo.getRewardType())
-                .bannerStartDate(vo.getBannerStartDate())
-                .bannerEndDate(vo.getBannerEndDate())
-                .willShow(vo.isWillShow())
-                .startDate(vo.getStartDate())
-                .endDate(vo.getEndDate())
+    public static AddExhibitionRequestDto toDto(AddExhibitionRequestVo addExhibitionRequestVo) {
+        return AddExhibitionRequestDto.builder()
+                .name(addExhibitionRequestVo.getName())
+                .description(addExhibitionRequestVo.getDescription())
+                .rewardType(addExhibitionRequestVo.getRewardType())
+                .bannerStartDate(addExhibitionRequestVo.getBannerStartDate())
+                .bannerEndDate(addExhibitionRequestVo.getBannerEndDate())
+                .willShow(addExhibitionRequestVo.isWillShow())
+                .startDate(addExhibitionRequestVo.getStartDate())
+                .endDate(addExhibitionRequestVo.getEndDate())
                 .build();
     }
 
     public Exhibition toEntity() {
         return Exhibition.builder()
-                .exhibitionId(exhibitionId)
                 .name(name)
                 .description(description)
                 .rewardType(rewardType)
